Refactor how cttm_options are send from PHP to JS
There were some issues with page builders firing the [travelersmap] shortcode multiple times on a single page load. This caused some weird JS double up variables like so:

 var cttm_options_params = etc..
 var cttm_shortcode_60b9be30c56fe = etc..
 var cttm_options_params = etc..
 var cttm_shortcode_60b9be30cfd2e = etc..

This refactor solves this by only rendering a single JS variable object like so:

 var var cttm_options_params = {"cttm_options":  (options here),  "cttm_shortcodes": [ (array of shortcodes here) ]

Now the JS only has a single global variable to inspect and we can remove all that dynamic window. variable stuff.

Also moved the shortcode init PHP into a class constructor as that's my personal preference for easy scope management, especially when trying to share data between multiple WP callbacks (such as the shortcode and wp footer callback).

Also removed a bunch of the JSON encode and decode stuff, as WordPress should handle that for us out of the box and give a correct JSON object for the browser to parse.