Extracting cttmMapLoop() to a different file
The cttmMapLoop() function was defined inside the initTravelersMap()
function and modified variables that were not defined inside
cttmMapLoop().

For clarity, this patch extracts the function to a
different file, and moves those variables back into cttmMapLoop (json_cttm_options and
mapindex), since they are only used inside it.

This helps make sure the functions are not entertwined.