Fixed wp_localize_script variables in Full Site Editing Themes
The map was not working in FSE themes introduced in WP5.9.
wp_localize_script was not adding the necessary variables to the front-end when a shortcode was inserted.
The registering of the script had to be moved to the shortcode function for the wp_localize_script to work (only in FSE themes)